+++ src/main/scala/farseek/core/package.scala
@@ -10,10 +10,12 @@ package object core {
   var gameDir: File = _
   var isDev: Boolean = _
 
-  lazy val spongeLoaded = try {
-    Class.forName("org.spongepowered.common.interfaces.world.IMixinWorldServer")
+  lazy val spongeLoaded = classLoaded("org.spongepowered.common.interfaces.world.IMixinWorldServer")
+
+  def classLoaded(name: String): Boolean =  try {
+    Class.forName(name)
     true
-  } catch { case e: ClassNotFoundException => false }
+  } catch { case _: ClassNotFoundException => false }
 
   final def allFiles(path: File): Array[File] =
     if(path == null || !path.exists || !path.canRead) Array()