+++ src/main/scala/farseek/world/gen/structure/StructureGenerationChunkProvider.scala
@@ -4,9 +4,7 @@ import farseek.core._
 import farseek.util.ImplicitConversions._
 import farseek.util.Reflection._
 import farseek.util._
-import farseek.world._
 import net.minecraft.world._
-import net.minecraft.world.chunk._
 import net.minecraft.world.gen.structure.MapGenStructure
 import net.minecraftforge.common.MinecraftForge._
 import net.minecraftforge.event.world.WorldEvent
@@ -22,7 +20,7 @@ import scala.collection.mutable
   *
   * @author delvr
   */
-class StructureGenerationChunkProvider(world: WorldServer) extends ServerBlockAccess with ChunkAccess with Logging {
+class StructureGenerationChunkProvider(world: WorldServer) extends Logging {
 
     val worldProvider = world.provider
 
@@ -38,26 +36,13 @@ class StructureGenerationChunkProvider(world: WorldServer) extends ServerBlockAc
         copyGenerator
     }
 
-    private val loadedChunks = mutable.Map[XZ, Chunk]()
-
     classFieldValues[MapGenStructure](generator).foreach(_.range = -1) // Disable structure generators
     EVENT_BUS.register(this)
 
-    def chunkAt(x: Int, z: Int) = {
-        val xChunk = x >> 4
-        val zChunk = z >> 4
-        loadedChunks.getOrElseUpdate((xChunk, zChunk), generateChunk(xChunk, zChunk))
-    }
-
-    private def generateChunk(xChunk: Int, zChunk: Int) = {
+    def generateChunk(xChunk: Int, zChunk: Int) = {
         val chunk = generator.provideChunk(xChunk, zChunk)
         chunk.setTerrainPopulated(true)
         chunk
-    }
-
-    // "real" chunk is ready with all structures in range, so we don't go there again
-    def onChunkProvided(chunk: Chunk) {
-        loadedChunks.remove(chunk.xPosition, chunk.zPosition)
     }
 
     @SubscribeEvent def onWorldUnload(event: WorldEvent.Unload) {
@@ -79,6 +64,4 @@ object StructureGenerationChunkProvider extends Logging {
     def apply(world: WorldServer) = providers.getOrElseUpdate(world, new StructureGenerationChunkProvider(world))
 
     def remove(worldProvider: WorldProvider) = providers.remove(worldProvider)
-
-    def onChunkProvided(chunk: Chunk) = providers.get(chunk.getWorld.provider).foreach(_.onChunkProvided(chunk))
 }