+++ src/main/scala/farseek/world/package.scala
@@ -4,6 +4,7 @@ import farseek.block._
 import farseek.util.ImplicitConversions._
 import farseek.util.{XYZ, _}
 import farseek.world.biome._
+import farseek.world.gen._
 import farseek.world.{AbsoluteCoordinates, CoordinateSystem}
 import net.minecraft.block.state._
 import net.minecraft.block.{BlockFalling, _}
@@ -32,7 +33,7 @@ package object world {
     val EndDimensionId = 1
 
     /**True while the game is generating or decorating chunks. */
-    var populating = false
+    def populating = BlockFalling.fallInstantly || populatingExtras
 
     // ----------------------------------------------------------------------------------------------------------------
     // Y-ranges