+++ src/main/scala/farseek/client/ConfigGui.scala
@@ -1,20 +1,18 @@
 package farseek.client
 
 import net.minecraft.client.Minecraft
+import net.minecraft.client.gui.GuiScreen
 import net.minecraftforge.fml.client.IModGuiFactory
-import net.minecraftforge.fml.client.IModGuiFactory.RuntimeOptionCategoryElement
 
 /** An [[IModGuiFactory]] implementation that defines a mod's [[RootConfigScreen]].
   * @author delvr
   */
 class ConfigGui extends IModGuiFactory {
 
-    def initialize(minecraft: Minecraft) {}
+  val hasConfigGui = true
+  val runtimeGuiCategories = null
 
-    def mainConfigGuiClass = classOf[RootConfigScreen]
+  def createConfigGui(parentScreen: GuiScreen) = new RootConfigScreen(parentScreen)
 
-    def runtimeGuiCategories = null
-
-    def getHandlerFor(element: RuntimeOptionCategoryElement) = null
+  def initialize(minecraft: Minecraft) {}
 }
-